requirements.txt
altgraph==0.17.2
attrs==21.4.0
cffi==1.15.1
iniconfig==1.1.1
MouseInfo==0.1.3
munch==2.5.0
numpy==1.23.1
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
proxy-tools==0.1.0
py==1.11.0
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
pygifsicle==1.0.5
pyinstaller==5.2
pyinstaller-hooks-contrib==2022.8
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
PyRect==0.2.0
PyScreeze==0.1.28
pytest==7.1.2
python3-xlib==0.15
pytweening==1.0.4
pywebview==3.6.3
QtPy==2.1.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
tomli==2.0.1
vext==0.7.6
vext.gi==0.7.4
webp==0.1.4
